lambda/patient-email-sender/requirements.txt
boto3==1.34.0
psycopg2-binary
requests==2.31.0

